+++ work/task4.c
@@ -3,18 +3,18 @@
 #include <stdlib.h>
 #include "task4.h"
 
-ll sumC(char *arr, int len)
+long long sumC(char *arr, int len)
 {
-	ll sum = 0;
+	long long sum = 0;
 	for (int i = 0; i < len; i++)
-		sum += (ll)(arr[i]);
+		sum += (long long)(arr[i]);
 	return sum;
 }
 
-ll sumR(char *arr, int len)
+long long sumR(char *arr, int len)
 {
 	if (len == 1)
-		return (ll)(arr[0]);
+		return (long long)(arr[0]);
 	else
-		return (ll)(arr[len - 1]) + sumR(arr, len - 1);
+		return (long long)(arr[len - 1]) + sumR(arr, len - 1);
 }
